+++ packages/nameof/src/tests/Transformation/TypeScriptAdapter.test.ts
@@ -51,6 +51,20 @@ export function TypeScriptAdapterTests(): void
                  * The item to check.
                  *
                  * @returns
+                 * A value indicating whether the specified {@linkcode item} is an accessor expression.
+                 */
+                public override IsAccessExpression(item: typescript.Node): boolean
+                {
+                    return super.IsAccessExpression(item);
+                }
+
+                /**
+                 * @inheritdoc
+                 *
+                 * @param item
+                 * The item to check.
+                 *
+                 * @returns
                  * A value indicating whether the specified {@linkcode item} is a string literal.
                  */
                 public override IsStringLiteral(item: typescript.Node): item is typescript.StringLiteral
@@ -271,6 +285,22 @@ export function TypeScriptAdapterTests(): void
                                         ts.factory.createIdentifier("nameof"),
                                         [],
                                         [])));
+                        });
+                });
+
+            suite(
+                nameOf<TestAdapter>((adapter) => adapter.IsAccessExpression),
+                () =>
+                {
+                    test(
+                        "Checking whether access expressions are detected properly…",
+                        () =>
+                        {
+                            let root = ts.factory.createIdentifier("here");
+
+                            ok(!adapter.IsAccessExpression(ts.factory.createNumericLiteral(420)));
+                            ok(adapter.IsAccessExpression(ts.factory.createPropertyAccessExpression(root, "be")));
+                            ok(adapter.IsAccessExpression(ts.factory.createElementAccessExpression(root, ts.factory.createStringLiteral("code"))));
                         });
                 });
 
@@ -585,19 +615,19 @@ export function TypeScriptAdapterTests(): void
                                 strictEqual(result.Type, NodeKind.IndexAccessNode);
                                 strictEqual(result.Source, node);
                                 deepStrictEqual(result.Expression, adapter.ParseInternal(node.expression, context));
-                                deepStrictEqual(result.Index, adapter.ParseInternal(node.argumentExpression, context));
+                                deepStrictEqual(result.Property, adapter.ParseInternal(node.argumentExpression, context));
 
                                 if (typeof index === "number")
                                 {
-                                    strictEqual(result.Index.Type, NodeKind.NumericLiteralNode);
-                                    strictEqual(result.Index.Source, node.argumentExpression);
-                                    strictEqual(result.Index.Value, index);
+                                    strictEqual(result.Property.Type, NodeKind.NumericLiteralNode);
+                                    strictEqual(result.Property.Source, node.argumentExpression);
+                                    strictEqual(result.Property.Value, index);
                                 }
                                 else
                                 {
-                                    strictEqual(result.Index.Type, NodeKind.StringLiteralNode);
-                                    strictEqual(result.Index.Source, index);
-                                    strictEqual(result.Index.Text, index.text);
+                                    strictEqual(result.Property.Type, NodeKind.StringLiteralNode);
+                                    strictEqual(result.Property.Source, index);
+                                    strictEqual(result.Property.Text, index.text);
                                 }
                             }
                         });
@@ -617,7 +647,7 @@ export function TypeScriptAdapterTests(): void
                             strictEqual(result.Type, NodeKind.IndexAccessNode);
                             strictEqual(result.Source, node);
                             deepStrictEqual(result.Expression, adapter.ParseInternal(node.objectType, context));
-                            deepStrictEqual(result.Index, adapter.ParseInternal(node.indexType, context));
+                            deepStrictEqual(result.Property, adapter.ParseInternal(node.indexType, context));
                         });
 
                     test(