+++ packages/webpack-workspace/.eslintrc.cjs
@@ -4,8 +4,7 @@ module.exports = {
     parserOptions: {
         project: [
             join(__dirname, "tsconfig.app.json"),
-            join(__dirname, "tsconfig.eslint.json"),
-            join(__dirname, "type-tests", "tsconfig.json")
+            join(__dirname, "tsconfig.editor.json")
         ]
     }
 };